mbuffer.h: Increase the buffer size for ARM EABI
struct ph_info_ch and struct ph_info has holes on ARM EABI
and the current size is too small for a LCR/mISDN loop
setup with 20 channels. Increase the limit a bit. The best
practice would be to either pack structures that are exposed
to userspace and to extract the size from the structures.