Export Q.931 state info in messages
For some services the apllication need the state info.
Exporting it in every message from layer3 is a safe option.
A status call would need locking (because a other thread could try to
free the process, so this is a no go).
TODO: test feature so applications can check if this is supported by the library.

Signed-off-by: Karsten Keil <kkeil@linux-pingi.de>